+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## UNRELEASED
+
+- Add support for deposit groups to group lifecycle report.
+
 ## 2024-12-03 v1.6.0
 
 - Add revision statistics to the group lifecycle report. Size of research group in this report